ASGI Support for HTTP (#11)
ASGI Support for HTTP events. We now are able to support frameworks like FastAPI to run natively using this plugin.
For that we created a new AsgiApp object on the C side that stores a reference to the ASGI coroutine entrypoint for the app. On each request we create a new AsgiEvent object that last for all the lifetime of the request and handles the state.
Introduced new interfacing functions `asgi_*` to synchronize the state of the request between Go and C.